- Move all encryption to module file - Cleanup imports - Added docstrings
