--- encryption/__init__.py
@@ -1,1 +1,0 @@
-from .client_side_encryption import ClientSideEncryption
